refactor: create Form\Template\Utils\{Frontend|Admin} class
