+++ src/Helpers/Form/Template/Utils/Admin.php
@@ -1,5 +1,5 @@
 <?php
-namespace Give\Helpers\Form\Template\Utils\Admin;
+namespace Give\Helpers\Form\Template\Utils;
 
 use Give\Form\Template;
 use Give\FormAPI\Form\Field;
@@ -8,57 +8,59 @@ use WP_Post;
 use Give\Helpers\Form\Template as FormTemplateUtils;
 
 
-/**
- * Render template setting in form metabox.
- *
- * @since 2.7.0
- *
- * @global WP_Post $post
- * @param Template $template
- * @return string
- */
-function renderMetaboxSettings( $template ) {
-	global $post;
+class Admin {
+	/**
+	 * Render template setting in form metabox.
+	 *
+	 * @since 2.7.0
+	 *
+	 * @global WP_Post $post
+	 * @param Template $template
+	 * @return string
+	 */
+	public static function renderMetaboxSettings( $template ) {
+		global $post;
 
-	ob_start();
+		ob_start();
 
-	$saveOptions = FormTemplateUtils::getOptions( $post->ID, $template->getID() );
+		$saveOptions = FormTemplateUtils::getOptions( $post->ID, $template->getID() );
 
-	/* @var Group $option */
-	foreach ( $template->getOptions()->groups as $group ) {
-		printf(
-			'<div class="give-row %1$s">',
-			$group->id
-		);
+		/* @var Group $option */
+		foreach ( $template->getOptions()->groups as $group ) {
+			printf(
+				'<div class="give-row %1$s">',
+				$group->id
+			);
 
-		printf(
-			'<div class="give-row-head">
+			printf(
+				'<div class="give-row-head">
 							<button type="button" class="handlediv" aria-expanded="true">
 								<span class="toggle-indicator"/>
 							</button>
 							<h2 class="hndle"><span>%1$s</span></h2>
 						</div>',
-			$group->name
-		);
+				$group->name
+			);
 
-		echo '<div class="give-row-body">';
+			echo '<div class="give-row-body">';
 
-		/* @var Field $field */
-		foreach ( $group->fields as $field ) {
-			$field = $field->toArray();
-			if ( isset( $saveOptions[ $group->id ][ $field['id'] ] ) ) {
-				$field['attributes']['value'] = $saveOptions[ $group->id ][ $field['id'] ];
+			/* @var Field $field */
+			foreach ( $group->fields as $field ) {
+				$field = $field->toArray();
+				if ( isset( $saveOptions[ $group->id ][ $field['id'] ] ) ) {
+					$field['attributes']['value'] = $saveOptions[ $group->id ][ $field['id'] ];
+				}
+
+				$field['id'] = "{$template->getID()}[{$group->id}][{$field['id']}]";
+
+				give_render_field( $field );
 			}
 
-			$field['id'] = "{$template->getID()}[{$group->id}][{$field['id']}]";
-
-			give_render_field( $field );
+			echo '</div></div>';
 		}
 
-		echo '</div></div>';
+		return ob_get_clean();
 	}
-
-	return ob_get_clean();
 }
