+++ src/Views/Form/Templates/Sequoia/sections/introduction.php
@@ -1,13 +1,13 @@
 <?php
 
-use function Give\Helpers\Form\Template\Utils\Frontend\getFormId;
+use Give\Helpers\Form\Template\Utils\Frontend as FrontendFormTemplateUtils;
 
-$formInfo = get_post( getFormId() );
+$formInfo = get_post( FrontendFormTemplateUtils::getFormId() );
 
 // Get headline and description
 $headline    = ! empty( $this->templateOptions['introduction']['headline'] ) ? $this->templateOptions['introduction']['headline'] : $formInfo->post_title;
 $description = ! empty( $this->templateOptions['introduction']['description'] ) ? $this->templateOptions['introduction']['description'] : $formInfo->post_excerpt;
-$image       = ! empty( $this->templateOptions['introduction']['image'] ) ? $this->templateOptions['introduction']['image'] : get_the_post_thumbnail_url( getFormId() );
+$image       = ! empty( $this->templateOptions['introduction']['image'] ) ? $this->templateOptions['introduction']['image'] : get_the_post_thumbnail_url( FrontendFormTemplateUtils::getFormId() );
 ?>
 
 <div class="give-section introduction">
