+++ src/Views/Form/Templates/Sequoia/views/receipt.php
@@ -2,7 +2,7 @@
 
 use Give\Views\IframeView;
 use Give\Helpers\Form\Template as FormTemplateUtils;
-use function Give\Helpers\Form\Template\Utils\Frontend\getPaymentId;
+use Give\Helpers\Form\Template\Utils\Frontend as FrontendFormTemplateUtils;
 use function give_get_gateway_admin_label as getGatewayLabel;
 use function give_currency_filter as filterCurrency;
 use function give_sanitize_amount as sanitizeAmount;
@@ -10,7 +10,7 @@ use function give_do_email_tags as formatContent;
 use Give_Payment as Payment;
 
 /* @var Payment $payment */
-$payment = new Payment( getPaymentId() );
+$payment = new Payment( FrontendFormTemplateUtils::getPaymentId() );
 $options = FormTemplateUtils::getOptions();
 
 ob_start();
